Fix PR 179 CI lint failures
